change to nameof() and throw argumentnull

+++ src/SlowCheetah.Tests/BuildTests/ConfigTransformTestsBase.cs
@@ -99,7 +99,10 @@ namespace SlowCheetah.Tests.BuildTests
         /// </summary>
         public void Dispose()
         {
-            Directory.Delete(this.OutputPath, true);
+            if (Directory.Exists(this.OutputPath))
+            {
+                Directory.Delete(this.OutputPath, true);
+            }
         }
     }
 }
